feat: refactor cluster options handling and remove unused code
